requirements.txt
chromadb==0.4.21
colorama==0.4.6
duckduckgo_search==5.3.1b1
huggingface_hub==0.20.3
newspaper4k==0.1.0.7
python-dotenv==1.0.1
Requests==2.32.3
setuptools==67.8.0
setuptools==63.2.0
torch==2.1.2+cu121
torch_xla==2.3.0
tqdm==4.65.0
transformers==4.36.2